feat: add expectations for exact match (#6)
This PR introduces a new set of extension methods for verifying the number of times a mocked interaction occurred using Mockolate, and removes the previous dummy extension. The changes add robust, readable assertions for checking if a method was called exactly, never, once, or twice, and include thorough test coverage for these scenarios.

Also add support for .NET 10.